Component keys are not stored in a static variable on each component class.
As long as the hash function is constexpr, there won't be any performance issue.

+++ hatcher/srcs/hatcher/constexpr_hash.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-template <typename T, size_t N = sizeof(T)>
+template <typename T, int N = sizeof(T)>
 constexpr unsigned int constexpr_hash(const T str);
 
 #include "constexpr_hash.ipp"
